## Related issue(s) and PR(s)
This PR closes #1976 .

## Description

Removed join
```
LEFT JOIN (SELECT file_id, (ARRAY_AGG(event ORDER BY started_at DESC))[1] AS event FROM sda.file_event_log GROUP BY file_id) log ON files.id = log.file_id
``` 
as no value from join was being returned (was no longer needed as of [commit](https://github.com/neicnordic/sensitive-data-archive/commit/dfc65bfb44cddbfa3e4db3699fa3dbab5ce06d88#diff-be0b1bf0b2a0b13034888f18466c7ad0450b170eaf7970e9c188fb2574ad0729))

Replaced join on view
```
LEFT JOIN local_ega.files lef ON files.stable_id = lef.stable_id` with `LEFT JOIN sda.checksums sha_arch ON files.id = sha_arch.file_id AND sha_arch.source = 'ARCHIVED'` to not have to evaluate the `local_ega.files
``` 
and replaced it with join to extract the relevant info from sda.checksums as per https://github.com/neicnordic/sensitive-data-archive/blob/main/postgresql/initdb.d/03.1_legacy_main.sql#L110

Also includes the removal of the following fields from the entries in the response on the `/metadata/datasets/*dataset/files` API to reduce amount of joins needed to improve performance of query:

* datasetId
* encryptedFileSize
* encryptedFileChecksum
* encryptedFileChecksumType

meaning the remaining fields in an entry in the response is:

* fileId
* displayFileName
* filePath
* decryptedFileSize
* decryptedFileChecksum
* decryptedFileChecksumType

Created a new struct for this response such to not affect the other APIs returning the `FileInfo` struct

Comparing queries in bp sda-staging:

### Old query:

#### EXPLAIN ANALYSE
```
Nested Loop Left Join  (cost=2.45..993.51 rows=1 width=392) (actual time=3.389..71.698 rows=5 loops=1)
  ->  Nested Loop Left Join  (cost=2.17..993.16 rows=1 width=307) (actual time=3.347..71.607 rows=5 loops=1)
        ->  Nested Loop  (cost=0.57..19.45 rows=1 width=238) (actual time=0.235..0.561 rows=5 loops=1)
              ->  Nested Loop  (cost=0.16..11.96 rows=1 width=48) (actual time=0.159..0.317 rows=5 loops=1)
                    ->  Seq Scan on file_dataset  (cost=0.00..4.18 rows=218 width=20) (actual time=0.021..0.050 rows=218 loops=1)
                    ->  Memoize  (cost=0.16..0.40 rows=1 width=36) (actual time=0.001..0.001 rows=0 loops=218)
                          Cache Key: file_dataset.dataset_id
                          Cache Mode: logical
                          Hits: 212  Misses: 6  Evictions: 0  Overflows: 0  Memory Usage: 1kB
                          ->  Index Scan using datasets_pkey on datasets  (cost=0.15..0.39 rows=1 width=36) (actual time=0.011..0.011 rows=0 loops=6)
                                Index Cond: (id = file_dataset.dataset_id)
                                Filter: (stable_id = 'aa-Dataset-jzf76d-deazcf'::text)
                                Rows Removed by Filter: 1
              ->  Index Scan using files_pkey on files  (cost=0.41..7.50 rows=1 width=206) (actual time=0.042..0.042 rows=1 loops=5)
                    Index Cond: (id = file_dataset.file_id)
        ->  Nested Loop Left Join  (cost=1.60..973.70 rows=1 width=91) (actual time=2.029..14.205 rows=1 loops=5)
              ->  Nested Loop Left Join  (cost=1.32..973.37 rows=1 width=107) (actual time=2.021..14.193 rows=1 loops=5)
                    ->  Nested Loop Left Join  (cost=1.04..973.03 rows=1 width=38) (actual time=2.010..14.181 rows=1 loops=5)
                          ->  Hash Join  (cost=0.76..972.70 rows=1 width=38) (actual time=1.961..14.131 rows=1 loops=5)
                                Hash Cond: (m.files_id = f.id)
                                ->  Seq Scan on main_to_files m  (cost=0.00..840.33 rows=50133 width=16) (actual time=0.005..6.261 rows=52209 loops=5)
                                ->  Hash  (cost=0.75..0.75 rows=1 width=38) (actual time=0.019..0.019 rows=1 loops=5)
                                      Buckets: 1024  Batches: 1  Memory Usage: 9kB
                                      ->  Index Scan using files_stable_id_key on files f  (cost=0.29..0.75 rows=1 width=38) (actual time=0.011..0.012 rows=1 loops=5)
                                            Index Cond: (stable_id = files.stable_id)
                          ->  Index Only Scan using unique_checksum on checksums  (cost=0.28..0.32 rows=1 width=16) (actual time=0.033..0.033 rows=1 loops=5)
                                Index Cond: ((file_id = f.id) AND (source = 'UPLOADED'::sda.checksum_source))
                                Heap Fetches: 0
                    ->  Index Scan using unique_checksum on checksums checksums_1  (cost=0.28..0.33 rows=1 width=85) (actual time=0.007..0.007 rows=1 loops=5)
                          Index Cond: ((file_id = f.id) AND (source = 'ARCHIVED'::sda.checksum_source))
              ->  Index Only Scan using unique_checksum on checksums checksums_2  (cost=0.28..0.32 rows=1 width=16) (actual time=0.005..0.005 rows=1 loops=5)
                    Index Cond: ((file_id = f.id) AND (source = 'UNENCRYPTED'::sda.checksum_source))
                    Heap Fetches: 0
  ->  Index Scan using unique_checksum on checksums checksums_3  (cost=0.28..0.33 rows=1 width=85) (actual time=0.005..0.006 rows=1 loops=5)
        Index Cond: ((file_id = files.id) AND (source = 'UNENCRYPTED'::sda.checksum_source))
Planning Time: 7.033 ms
Execution Time: 72.248 ms
``` 

#### Output (tested with a test dataset `aa-Dataset-jzf76d-deazcf` in bp sda-staging)
```
aa-File-8dfk3p-zqxjg2,aa-Dataset-jzf76d-deazcf,pan_test_file_2.c4gh,2aa6fa292e28d76df0abb008da4c3126eac4753c@lifescience-ri.eu,DATASET_pantest000001/pan_test_file_2.c4gh,1049024,686acd9b42c019a2486ec9fa670dd9fae7e8783f28898169a3ce962b6fe90735,SHA256,1048576,30e14955ebf1352266dc2ff8067e68104607e750abb9d3b36582b8af909fcb58,SHA256
aa-File-g5xucu-c2s27n,aa-Dataset-jzf76d-deazcf,pan_test_file_4.c4gh,2aa6fa292e28d76df0abb008da4c3126eac4753c@lifescience-ri.eu,DATASET_pantest000001/pan_test_file_4.c4gh,1049024,43f5db410e4897e70c7380e5d255b856f1bb0d6473dbeaac2f2b97ff44c6be25,SHA256,1048576,30e14955ebf1352266dc2ff8067e68104607e750abb9d3b36582b8af909fcb58,SHA256
aa-File-tty3es-dpk2qv,aa-Dataset-jzf76d-deazcf,pan_test_file_1.c4gh,2aa6fa292e28d76df0abb008da4c3126eac4753c@lifescience-ri.eu,DATASET_pantest000001/pan_test_file_1.c4gh,1049024,951e9b56713c2cddcf3991c8e205348f3aeaec8872b8797bd4dbd6b93bd52b31,SHA256,1048576,30e14955ebf1352266dc2ff8067e68104607e750abb9d3b36582b8af909fcb58,SHA256
aa-File-6349su-k8fphk,aa-Dataset-jzf76d-deazcf,pan_test_file_5.c4gh,2aa6fa292e28d76df0abb008da4c3126eac4753c@lifescience-ri.eu,DATASET_pantest000001/pan_test_file_5.c4gh,1049024,a445e370d48519126fd3df1c68ce9b5b97c37d1a2cff1175d9d445ca9695f7ae,SHA256,1048576,30e14955ebf1352266dc2ff8067e68104607e750abb9d3b36582b8af909fcb58,SHA256
aa-File-6zqpct-b9c32e,aa-Dataset-jzf76d-deazcf,pan_test_file_3.c4gh,2aa6fa292e28d76df0abb008da4c3126eac4753c@lifescience-ri.eu,DATASET_pantest000001/pan_test_file_3.c4gh,1049024,6da15fa01ebdef15f2a242baa028af02b3634177119a29e18d36cc7e7bcc5253,SHA256,1048576,30e14955ebf1352266dc2ff8067e68104607e750abb9d3b36582b8af909fcb58,SHA256
```

### New query:
#### EXPLAIN ANALYSE
```
Nested Loop Left Join  (cost=0.85..19.80 rows=1 width=315) (actual time=0.205..0.410 rows=5 loops=1)
  ->  Nested Loop  (cost=0.57..19.45 rows=1 width=230) (actual time=0.185..0.351 rows=5 loops=1)
        ->  Nested Loop  (cost=0.16..11.96 rows=1 width=48) (actual time=0.171..0.303 rows=5 loops=1)
              ->  Seq Scan on file_dataset  (cost=0.00..4.18 rows=218 width=20) (actual time=0.019..0.064 rows=218 loops=1)
              ->  Memoize  (cost=0.16..0.40 rows=1 width=36) (actual time=0.001..0.001 rows=0 loops=218)
                    Cache Key: file_dataset.dataset_id
                    Cache Mode: logical
                    Hits: 212  Misses: 6  Evictions: 0  Overflows: 0  Memory Usage: 1kB
                    ->  Index Scan using datasets_pkey on datasets  (cost=0.15..0.39 rows=1 width=36) (actual time=0.006..0.006 rows=0 loops=6)
                          Index Cond: (id = file_dataset.dataset_id)
                          Filter: (stable_id = 'aa-Dataset-jzf76d-deazcf'::text)
                          Rows Removed by Filter: 1
        ->  Index Scan using files_pkey on files  (cost=0.41..7.50 rows=1 width=198) (actual time=0.008..0.008 rows=1 loops=5)
              Index Cond: (id = file_dataset.file_id)
  ->  Index Scan using unique_checksum on checksums sha_unenc  (cost=0.28..0.33 rows=1 width=85) (actual time=0.006..0.007 rows=1 loops=5)
        Index Cond: ((file_id = files.id) AND (source = 'UNENCRYPTED'::sda.checksum_source))
Planning Time: 1.460 ms
Execution Time: 0.513 ms
``` 

#### Output (tested with a test dataset `aa-Dataset-jzf76d-deazcf` in bp sda-staging)
```aa-File-8dfk3p-zqxjg2,pan_test_file_2.c4gh,2aa6fa292e28d76df0abb008da4c3126eac4753c@lifescience-ri.eu,DATASET_pantest000001/pan_test_file_2.c4gh,1048576,30e14955ebf1352266dc2ff8067e68104607e750abb9d3b36582b8af909fcb58,SHA256
aa-File-g5xucu-c2s27n,pan_test_file_4.c4gh,2aa6fa292e28d76df0abb008da4c3126eac4753c@lifescience-ri.eu,DATASET_pantest000001/pan_test_file_4.c4gh,1048576,30e14955ebf1352266dc2ff8067e68104607e750abb9d3b36582b8af909fcb58,SHA256
aa-File-tty3es-dpk2qv,pan_test_file_1.c4gh,2aa6fa292e28d76df0abb008da4c3126eac4753c@lifescience-ri.eu,DATASET_pantest000001/pan_test_file_1.c4gh,1048576,30e14955ebf1352266dc2ff8067e68104607e750abb9d3b36582b8af909fcb58,SHA256
aa-File-6349su-k8fphk,pan_test_file_5.c4gh,2aa6fa292e28d76df0abb008da4c3126eac4753c@lifescience-ri.eu,DATASET_pantest000001/pan_test_file_5.c4gh,1048576,30e14955ebf1352266dc2ff8067e68104607e750abb9d3b36582b8af909fcb58,SHA256
aa-File-6zqpct-b9c32e,pan_test_file_3.c4gh,2aa6fa292e28d76df0abb008da4c3126eac4753c@lifescience-ri.eu,DATASET_pantest000001/pan_test_file_3.c4gh,1048576,30e14955ebf1352266dc2ff8067e68104607e750abb9d3b36582b8af909fcb58,SHA256
```

And remove TestGetFiles unit test which just verifies which query is ran by the `getFiles` func

## How to test
Verify that integration tests still pass