@KarlG-nbis 2025-09-25
> Impressive performance improvement. Great work! I can approve the PR once we get confirmed about the removal of 4 items from the `metadata/datasets/*dataset/files` endpoint.
> 
> Also a minor thing: The comment `// FileInfo is returned by the metadata endpoin` for the `FileInfo` struct can be removed since all items of the `FileInfo` are not returned by the metadata endpoint anymore.

Yeah, updated now to not introduce the `DatasetFile struct` and instead modify the `FileInfo struct` and just removed the getDatasetFileInfo` to be replaced by `GetDatasetFileStableId` as usage was just using the `FileInfo.FileId`

Felt cleaner to do so in this PR than require another PR to just update the getDatasetFileInfo since otherwise would need this new DatasetFile struct temporarily until the next PR to deal with the `getDatasetFileInfo func`